sisteminha-main/backend/requirements.txt
asgiref==3.8.1
Django==4.2.21
django-cors-headers==4.6.0
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
Markdown==3.7
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.2
uritemplate==4.1.1
djangorestframework-simplejwt==5.2.2
pytest==8.4.1
mutpy==0.6.1